fix(frontend): remove fragile setTimeout loading logic and modularize Navbar
- I-04: Removed setTimeout-based loading messages in page.jsx
- I-06: Extracted DesktopNav, MobileMenu, UserMenu from monolithic Navbar
- Navbar reduced from ~540 lines to ~260 lines

Closes #20, Closes #22
